requirements.txt
Django
django-environ
djangorestframework
psycopg2
drf_yasg
